plotify-api/requirements.txt
click==8.1.7
fastapi==0.111.1
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
hatched==0.2.0
shapely<2.0
opencv-python-headless<4.0
numpy<2.0
vpype
toml
git+https://github.com/serycjon/vpype-flow-imager.git#egg=vpype-flow-imager[all]
inkex
svgutils
replicate